+++ pilot/pkg/model/xds_cache.go
@@ -15,392 +15,250 @@
 package model
 
 import (
-	"fmt"
-	"sync"
-	"time"
-
 	discovery "github.com/envoyproxy/go-control-plane/envoy/service/discovery/v3"
-	"github.com/google/go-cmp/cmp"
-	"github.com/hashicorp/golang-lru/v2/simplelru"
-	"golang.org/x/exp/slices"
-	"google.golang.org/protobuf/testing/protocmp"
 
 	"istio.io/istio/pilot/pkg/features"
+	"istio.io/istio/pkg/config/schema/kind"
 	"istio.io/istio/pkg/util/sets"
-	"istio.io/pkg/monitoring"
 )
 
-var enableStats = func() bool {
-	return features.EnableXDSCacheMetrics
+type XdsCacheImpl struct {
+	cds typedXdsCache[uint64]
+	eds typedXdsCache[uint64]
+	rds typedXdsCache[uint64]
+	sds typedXdsCache[string]
 }
 
-var (
-	xdsCacheReads = monitoring.RegisterIf(monitoring.NewSum(
-		"xds_cache_reads",
-		"Total number of xds cache xdsCacheReads.",
-		monitoring.WithLabels(typeTag)), enableStats)
-
-	xdsCacheEvictions = monitoring.RegisterIf(monitoring.NewSum(
-		"xds_cache_evictions",
-		"Total number of xds cache evictions.",
-		monitoring.WithLabels(typeTag)), enableStats)
-
-	xdsCacheSize = monitoring.RegisterIf(monitoring.NewGauge(
-		"xds_cache_size",
-		"Current size of xds cache"), enableStats)
-
-	dependentConfigSize = monitoring.RegisterIf(monitoring.NewGauge(
-		"xds_cache_dependent_config_size",
-		"Current size of dependent configs"), enableStats)
-
-	xdsCacheHits             = xdsCacheReads.With(typeTag.Value("hit"))
-	xdsCacheMisses           = xdsCacheReads.With(typeTag.Value("miss"))
-	xdsCacheEvictionsOnClear = xdsCacheEvictions.With(typeTag.Value("clear"))
-	xdsCacheEvictionsOnSize  = xdsCacheEvictions.With(typeTag.Value("size"))
-)
-
-func hit() {
-	xdsCacheHits.Increment()
-}
-
-func miss() {
-	xdsCacheMisses.Increment()
-}
-
-func size(cs int) {
-	xdsCacheSize.Record(float64(cs))
-}
-
-type CacheToken uint64
-
-type dependents interface {
-	DependentConfigs() []ConfigHash
-}
-
-// xdsCache interface defines a store for caching XDS responses.
+// XdsCache interface defines a store for caching XDS responses.
 // All operations are thread safe.
-type xdsCache[K comparable] interface {
+type XdsCache interface {
 	// Run starts a background thread periodically flush evicted indexes
 	Run(stop <-chan struct{})
 	// Add adds the given XdsCacheEntry with the value for the given pushContext to the cache.
 	// If the cache has been updated to a newer push context, the write will be dropped silently.
 	// This ensures stale data does not overwrite fresh data when dealing with concurrent
 	// writers.
-	Add(key K, entry dependents, pushRequest *PushRequest, value *discovery.Resource)
+	Add(entry XdsCacheEntry, pushRequest *PushRequest, value *discovery.Resource)
 	// Get retrieves the cached value if it exists.
-	Get(key K) *discovery.Resource
+	Get(entry XdsCacheEntry) *discovery.Resource
 	// GetWithRequest retrieves the cached value if it was generated by a newer push request.
-	GetWithRequest(key K, pushReq *PushRequest) *discovery.Resource
+	GetWithRequest(entry XdsCacheEntry, pushReq *PushRequest) *discovery.Resource
 	// Clear removes the cache entries that are dependent on the configs passed.
 	Clear(sets.Set[ConfigKey])
 	// ClearAll clears the entire cache.
 	ClearAll()
-	// Keys returns all currently configured keys. This is for testing/debug only
-	Keys() []K
-	// Snapshot returns a snapshot of all keys and values. This is for testing/debug only
+	// Keys returns all currently configured keys for the type. This is for testing/debug only
+	Keys(t string) []any
+	// Snapshot returns a snapshot of all values. This is for testing/debug only
 	Snapshot() []*discovery.Resource
 }
 
-// newXdsCache returns an instance of a cache.
-func newXdsCache[K comparable]() xdsCache[K] {
-	cache := &lruCache[K]{
-		enableAssertions: features.EnableUnsafeAssertions,
-		configIndex:      map[ConfigHash]sets.Set[K]{},
-		evictQueue:       make([]evictKeyConfigs[K], 0, 1000),
+// XdsCacheEntry interface defines functions that should be implemented by
+// resources that can be cached.
+type XdsCacheEntry interface {
+	Type() string
+	// Key is the key to be used in cache.
+	Key() any
+	// DependentConfigs is config items that this cache key is dependent on.
+	// Whenever these configs change, we should invalidate this cache entry.
+	DependentConfigs() []ConfigHash
+	// Cacheable indicates whether this entry is valid for cache. For example
+	// for EDS to be cacheable, the Endpoint should have corresponding service.
+	Cacheable() bool
+}
+
+const (
+	CDSType = "cds"
+	EDSType = "eds"
+	RDSType = "rds"
+	SDSType = "sds"
+)
+
+// NewXdsCache returns an instance of a cache.
+func NewXdsCache() XdsCache {
+	var cache = XdsCacheImpl{
+		eds: newTypedXdsCache[uint64](),
 	}
-	cache.store = newLru(cache.onEvict)
+	if features.EnableCDSCaching {
+		cache.cds = newTypedXdsCache[uint64]()
+	} else {
+		cache.cds = disabledCache[uint64]{}
+	}
+	if features.EnableRDSCaching {
+		cache.rds = newTypedXdsCache[uint64]()
+	} else {
+		cache.rds = disabledCache[uint64]{}
+	}
+
+	cache.sds = newTypedXdsCache[string]()
+
 	return cache
 }
 
-type evictKeyConfigs[K comparable] struct {
-	key              K
-	dependentConfigs []ConfigHash
+func (x XdsCacheImpl) Run(stop <-chan struct{}) {
+	x.cds.Run(stop)
+	x.eds.Run(stop)
+	x.rds.Run(stop)
+	x.sds.Run(stop)
 }
 
-type lruCache[K comparable] struct {
-	enableAssertions bool
-	store            simplelru.LRUCache[K, cacheValue]
-	// token stores the latest token of the store, used to prevent stale data overwrite.
-	// It is refreshed when Clear or ClearAll are called
-	token       CacheToken
-	mu          sync.RWMutex
-	configIndex map[ConfigHash]sets.Set[K]
-
-	evictQueue []evictKeyConfigs[K]
-
-	// mark whether a key is evicted on Clear call, passively.
-	evictedOnClear bool
-}
-
-var _ xdsCache[uint64] = &lruCache[uint64]{}
-
-func newLru[K comparable](evictCallback simplelru.EvictCallback[K, cacheValue]) simplelru.LRUCache[K, cacheValue] {
-	sz := features.XDSCacheMaxSize
-	if sz <= 0 {
-		sz = 20000
-	}
-	l, err := simplelru.NewLRU(sz, evictCallback)
-	if err != nil {
-		panic(fmt.Errorf("invalid lru configuration: %v", err))
-	}
-	return l
-}
-
-func (l *lruCache[K]) Run(stop <-chan struct{}) {
-	interval := features.XDSCacheIndexClearInterval
-	go func() {
-		ticker := time.NewTicker(interval)
-		defer ticker.Stop()
-		for {
-			select {
-			case <-ticker.C:
-				l.mu.Lock()
-				for _, keyConfigs := range l.evictQueue {
-					l.clearConfigIndex(keyConfigs.key, keyConfigs.dependentConfigs)
-				}
-				l.evictQueue = l.evictQueue[:0:1000]
-				l.recordDependentConfigSize()
-				l.mu.Unlock()
-			case <-stop:
-				return
-			}
-		}
-	}()
-}
-
-func (l *lruCache[K]) recordDependentConfigSize() {
-	if !enableStats() {
+func (x XdsCacheImpl) Add(entry XdsCacheEntry, pushRequest *PushRequest, value *discovery.Resource) {
+	if !entry.Cacheable() {
 		return
 	}
-	dsize := 0
-	for _, dependents := range l.configIndex {
-		dsize += len(dependents)
-	}
-	dependentConfigSize.Record(float64(dsize))
-}
-
-// This is the callback passed to LRU, it will be called whenever a key is removed.
-func (l *lruCache[K]) onEvict(k K, v cacheValue) {
-	if l.evictedOnClear {
-		xdsCacheEvictionsOnClear.Increment()
-	} else {
-		xdsCacheEvictionsOnSize.Increment()
-	}
-
-	// async clearing indexes
-	l.evictQueue = append(l.evictQueue, evictKeyConfigs[K]{k, v.dependentConfigs})
-}
-
-func (l *lruCache[K]) updateConfigIndex(k K, dependentConfigs []ConfigHash) {
-	for _, cfg := range dependentConfigs {
-		sets.InsertOrNew(l.configIndex, cfg, k)
+	k := entry.Key()
+	switch entry.Type() {
+	case CDSType:
+		key := k.(uint64)
+		x.cds.Add(key, entry, pushRequest, value)
+	case EDSType:
+		key := k.(uint64)
+		x.eds.Add(key, entry, pushRequest, value)
+	case SDSType:
+		key := k.(string)
+		x.sds.Add(key, entry, pushRequest, value)
+	case RDSType:
+		key := k.(uint64)
+		x.rds.Add(key, entry, pushRequest, value)
+	default:
+		log.Errorf("unknown type %s", entry.Type())
 	}
 }
 
-func (l *lruCache[K]) clearConfigIndex(k K, dependentConfigs []ConfigHash) {
-	c, exists := l.store.Get(k)
-	if exists {
-		newDependents := c.dependentConfigs
-		// we only need to clear configs {old difference new}
-		dependents := sets.New[ConfigHash](dependentConfigs...).Difference(sets.New[ConfigHash](newDependents...))
-		for cfg := range dependents {
-			sets.DeleteCleanupLast(l.configIndex, cfg, k)
-		}
-		return
-	}
-	for _, cfg := range dependentConfigs {
-		sets.DeleteCleanupLast(l.configIndex, cfg, k)
-	}
-}
-
-// assertUnchanged checks that a cache entry is not changed. This helps catch bad cache invalidation
-// We should never have a case where we overwrite an existing item with a new change. Instead, when
-// config sources change, Clear/ClearAll should be called. At this point, we may get multiple writes
-// because multiple writers may get cache misses concurrently, but they ought to generate identical
-// configuration. This also checks that our XDS config generation is deterministic, which is a very
-// important property.
-func (l *lruCache[K]) assertUnchanged(key K, existing *discovery.Resource, replacement *discovery.Resource) {
-	if l.enableAssertions {
-		if existing == nil {
-			// This is a new addition, not an update
-			return
-		}
-		// Record time so that we can correlate when the error actually happened, since the async reporting
-		// may be delayed
-		t0 := time.Now()
-		// This operation is really slow, which makes tests fail for unrelated reasons, so we process it async.
-		go func() {
-			if !cmp.Equal(existing, replacement, protocmp.Transform()) {
-				warning := fmt.Errorf("assertion failed at %v, cache entry changed but not cleared for key %v: %v\n%v\n%v",
-					t0, key, cmp.Diff(existing, replacement, protocmp.Transform()), existing, replacement)
-				panic(warning)
-			}
-		}()
-	}
-}
-
-func (l *lruCache[K]) Add(k K, entry dependents, pushReq *PushRequest, value *discovery.Resource) {
-	if pushReq == nil || pushReq.Start.Equal(time.Time{}) {
-		return
-	}
-	// It will not overflow until year 2262
-	token := CacheToken(pushReq.Start.UnixNano())
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	if token < l.token {
-		// entry may be stale, we need to drop it. This can happen when the cache is invalidated
-		// after we call Get.
-		return
-	}
-	cur, f := l.store.Get(k)
-	if f {
-		// This is the stale or same resource
-		if token <= cur.token {
-			return
-		}
-		if l.enableAssertions {
-			l.assertUnchanged(k, cur.value, value)
-		}
-	}
-
-	dependentConfigs := entry.DependentConfigs()
-	toWrite := cacheValue{value: value, token: token, dependentConfigs: dependentConfigs}
-	l.store.Add(k, toWrite)
-	l.token = token
-	l.updateConfigIndex(k, dependentConfigs)
-
-	// we have to make sure we evict old entries with the same key
-	// to prevent leaking in the index maps
-	if f {
-		l.evictQueue = append(l.evictQueue, evictKeyConfigs[K]{k, cur.dependentConfigs})
-	}
-	size(l.store.Len())
-}
-
-type cacheValue struct {
-	value            *discovery.Resource
-	token            CacheToken
-	dependentConfigs []ConfigHash
-}
-
-func (l *lruCache[K]) Get(key K) *discovery.Resource {
-	return l.get(key, 0)
-}
-
-// GetWithRequest return the cached value generated by newer push request than the given one.
-func (l *lruCache[K]) GetWithRequest(key K, pushReq *PushRequest) *discovery.Resource {
-	return l.get(key, CacheToken(pushReq.Start.UnixNano()))
-}
-
-// get return the cached value if it exists.
-func (l *lruCache[K]) get(key K, token CacheToken) *discovery.Resource {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	cv, ok := l.store.Get(key)
-	if !ok || cv.value == nil {
-		miss()
+func (x XdsCacheImpl) Get(entry XdsCacheEntry) *discovery.Resource {
+	if !entry.Cacheable() {
 		return nil
 	}
-	if cv.token >= token {
-		hit()
-		return cv.value
+
+	k := entry.Key()
+	switch entry.Type() {
+	case CDSType:
+		key := k.(uint64)
+		return x.cds.Get(key)
+	case EDSType:
+		key := k.(uint64)
+		return x.eds.Get(key)
+	case SDSType:
+		key := k.(string)
+		return x.sds.Get(key)
+	case RDSType:
+		key := k.(uint64)
+		return x.rds.Get(key)
+	default:
+		log.Errorf("unknown type %s", entry.Type())
+		return nil
 	}
-	miss()
+}
+
+func (x XdsCacheImpl) Clear(s sets.Set[ConfigKey]) {
+	x.cds.Clear(s)
+	// clear all EDS cache for PA change
+	if HasConfigsOfKind(s, kind.PeerAuthentication) {
+		x.eds.ClearAll()
+	} else {
+		x.eds.Clear(s)
+	}
+	x.rds.Clear(s)
+	x.sds.Clear(s)
+}
+
+func (x XdsCacheImpl) ClearAll() {
+	x.cds.ClearAll()
+	x.eds.ClearAll()
+	x.rds.ClearAll()
+	x.sds.ClearAll()
+}
+
+func (x XdsCacheImpl) Keys(t string) []any {
+	switch t {
+	case CDSType:
+		keys := x.cds.Keys()
+		return convertToAnySlices(keys)
+	case EDSType:
+		keys := x.eds.Keys()
+		return convertToAnySlices(keys)
+	case SDSType:
+		keys := x.sds.Keys()
+		return convertToAnySlices(keys)
+	case RDSType:
+		keys := x.rds.Keys()
+		return convertToAnySlices(keys)
+	default:
+		return nil
+	}
+}
+
+func convertToAnySlices[K comparable](in []K) []any {
+	var out = make([]any, len(in))
+	for i, k := range in {
+		out[i] = k
+	}
+	return out
+}
+
+func (x XdsCacheImpl) Snapshot() []*discovery.Resource {
+	var out []*discovery.Resource
+	out = append(out, x.cds.Snapshot()...)
+	out = append(out, x.eds.Snapshot()...)
+	out = append(out, x.rds.Snapshot()...)
+	out = append(out, x.sds.Snapshot()...)
+	return out
+}
+
+func (x XdsCacheImpl) GetWithRequest(entry XdsCacheEntry, pushReq *PushRequest) *discovery.Resource {
+	if !entry.Cacheable() {
+		return nil
+	}
+
+	k := entry.Key()
+	switch entry.Type() {
+	case CDSType:
+		key := k.(uint64)
+		return x.cds.GetWithRequest(key, pushReq)
+	case EDSType:
+		key := k.(uint64)
+		return x.eds.GetWithRequest(key, pushReq)
+	case SDSType:
+		key := k.(string)
+		return x.sds.GetWithRequest(key, pushReq)
+	case RDSType:
+		key := k.(uint64)
+		return x.rds.GetWithRequest(key, pushReq)
+	default:
+		log.Errorf("unknown type %s", entry.Type())
+		return nil
+	}
+}
+
+// DisabledCache is a cache that is always empty
+type DisabledCache struct{}
+
+func (d DisabledCache) Run(stop <-chan struct{}) {
+}
+
+func (d DisabledCache) Add(entry XdsCacheEntry, pushRequest *PushRequest, value *discovery.Resource) {
+}
+
+func (d DisabledCache) Get(entry XdsCacheEntry) *discovery.Resource {
 	return nil
 }
 
-func (l *lruCache[K]) Clear(configs sets.Set[ConfigKey]) {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	l.token = CacheToken(time.Now().UnixNano())
-	l.evictedOnClear = true
-	defer func() {
-		l.evictedOnClear = false
-	}()
-	for ckey := range configs {
-		referenced := l.configIndex[ckey.HashCode()]
-		delete(l.configIndex, ckey.HashCode())
-		for key := range referenced {
-			l.store.Remove(key)
-		}
-	}
-	size(l.store.Len())
-}
-
-func (l *lruCache[K]) ClearAll() {
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	l.token = CacheToken(time.Now().UnixNano())
-	// Purge with an evict function would turn up to be pretty slow since
-	// it runs the function for every key in the store, might be better to just
-	// create a new store.
-	l.store = newLru(l.onEvict)
-	l.configIndex = map[ConfigHash]sets.Set[K]{}
-	l.evictQueue = l.evictQueue[:0:1000]
-	size(l.store.Len())
-}
-
-func (l *lruCache[K]) Keys() []K {
-	l.mu.RLock()
-	defer l.mu.RUnlock()
-	return slices.Clone(l.store.Keys())
-}
-
-func (l *lruCache[K]) Snapshot() []*discovery.Resource {
-	l.mu.RLock()
-	defer l.mu.RUnlock()
-	iKeys := l.store.Keys()
-	res := make([]*discovery.Resource, len(iKeys))
-	for i, ik := range iKeys {
-		v, ok := l.store.Get(ik)
-		if !ok {
-			continue
-		}
-
-		res[i] = v.value
-	}
-	return res
-}
-
-func (l *lruCache[K]) indexLength() int {
-	l.mu.RLock()
-	defer l.mu.RUnlock()
-	return len(l.configIndex)
-}
-
-func (l *lruCache[K]) configIndexSnapshot() map[ConfigHash]sets.Set[K] {
-	l.mu.RLock()
-	defer l.mu.RUnlock()
-	res := make(map[ConfigHash]sets.Set[K], len(l.configIndex))
-	for k, v := range l.configIndex {
-		res[k] = v
-	}
-	return res
-}
-
-// disabledCache is a cache that is always empty
-type disabledCache[K comparable] struct{}
-
-var _ xdsCache[uint64] = &disabledCache[uint64]{}
-
-func (d disabledCache[K]) Run(stop <-chan struct{}) {
-}
-
-func (d disabledCache[K]) Add(k K, entry dependents, pushReq *PushRequest, value *discovery.Resource) {
-}
-
-func (d disabledCache[K]) Get(k K) *discovery.Resource {
+func (d DisabledCache) GetWithRequest(entry XdsCacheEntry, pushReq *PushRequest) *discovery.Resource {
 	return nil
 }
 
-func (d disabledCache[K]) GetWithRequest(k K, pushReq *PushRequest) *discovery.Resource {
+func (d DisabledCache) Clear(s sets.Set[ConfigKey]) {
+}
+
+func (d DisabledCache) ClearAll() {
+}
+
+func (d DisabledCache) Keys(t string) []any {
 	return nil
 }
 
-func (d disabledCache[K]) Clear(configsUpdated sets.Set[ConfigKey]) {}
+func (d DisabledCache) Snapshot() []*discovery.Resource {
+	return nil
+}
 
-func (d disabledCache[K]) ClearAll() {}
-
-func (d disabledCache[K]) Keys() []K { return nil }
-
-func (d disabledCache[K]) Snapshot() []*discovery.Resource { return nil }
+var _ XdsCache = &DisabledCache{}
